+++ shared/src/main/kotlin/com/example/strata/shared/dto/Queries.kt
@@ -1,13 +1,13 @@
 package com.example.strata.shared.dto
 
-import java.math.BigDecimal
+import java.math.BigInteger
 import java.time.Instant
 import java.util.UUID
 
 data class AccountSummaryResponse(
     val accountId: UUID,
     val accountName: String,
-    val balance: BigDecimal,
+    val balance: BigInteger,
     val lastUpdated: Instant,
     val eventVersion: Long
 )
